% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_studiesinfo.R
\name{diag_studiesinfo}
\alias{diag_studiesinfo}
\title{diag_studiesinfo}
\usage{
diag_studiesinfo(tab, type)
}
\arguments{
\item{tab}{A table containing the studiesinfo.}

\item{type}{A character string. Either "discovery" or "validation".}
}
\description{
Helper function to import studiesinfos for diagnostic
  studies from table
}
\details{
The function takes a table as input and converts it to a
  studiesinfo list.
}
