% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_batch_correction.R
\name{geo_batch_correction}
\alias{geo_batch_correction}
\title{geo_batch_correction}
\usage{
geo_batch_correction(mergedset, batch, design, idtype)
}
\arguments{
\item{mergedset}{A large Expression Set. The output of the function
`geo_merge`. Please note, that mergedset holds data, which are not
yet batch corrected.}

\item{batch}{Takes the results from \code{create_batch()} as input.}

\item{design}{A two-column matrix. Holds information on the studie's
target variable and is created by `stats::model.matrix(~ diagnosis)`.}

\item{idtype}{A character string. The type of ID used to name the
genes. One of 'entrez' or 'affy' intended to use either entrez IDs or
affy IDs. Caution: when using entrez IDs, missing and duplicated IDs
are being removed!}
}
\description{
Helper function correcting for batch effects and mapping
  affy probes to Entrez IDs
}
\details{
This function takes a Bioconductor's ExpressionSet class (the
  output of the function `geo_merge`) and outputs a batch corrected
  matrix containing expression data. In order to correct for occurring
  batch effects and other unwanted variation in high-throughput
  experiments the `ComBat` function from the sva package is conducted.
  The affy probes are mapped to their Entrez IDs. Thereby, empty and
  replicated character strings are removed.
}
\references{
W.E. Johnson, C. Li, and A. Rabinovic. Adjusting batch effects
  in microarray data using empirical bayes methods. Biostatistics,
  8(1):118–127, 2007. Jeffrey T. Leek, W. Evan Johnson, Hilary S. Parker,
  Elana J. Fertig, Andrew E. Jaffe, John D. Storey, Yuqing Zhang and
  Leonardo Collado Torres (2019). sva: Surrogate Variable Analysis.
  R package version 3.30.1.
}
