% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_batch_detection.R
\name{geo_batch_detection}
\alias{geo_batch_detection}
\title{geo_batch_detection}
\usage{
geo_batch_detection(mergeset, batch)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes,
columns = samples).}

\item{batch}{Takes the results from \code{create_batch()} as input.}
}
\description{
Helper function to detect batches
}
