% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_create_batch.R
\name{geo_create_batch}
\alias{geo_create_batch}
\title{geo_create_batch}
\usage{
geo_create_batch(sample_metadata)
}
\arguments{
\item{sample_metadata}{A data frame. The data frame holding the
sample metadata.}
}
\description{
Helper function to create batch
}
