% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_create_diagnosis.R
\name{geo_create_diagnosis}
\alias{geo_create_diagnosis}
\title{geo_create_diagnosis}
\usage{
geo_create_diagnosis(vector, controlname, targetname)
}
\arguments{
\item{vector}{A vector of characters. The target variable, consisting of
two levels, named `controlname` and `targetname`.}

\item{controlname}{A character string. Name of the the controls,
specified in the 'target' column of `sample_metadata` (default: Control).}

\item{targetname}{A character string. Name of the the targets, specified
in the 'target' column of `sample_metadata` (default: Target).}
}
\description{
Helper function to transform the target variable into a
  binary numeric diagnosis variable.
}
