% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_create_diagnosisbatch.R
\name{geo_create_diagnosisbatch}
\alias{geo_create_diagnosisbatch}
\title{geo_create_diagnosisbatch}
\usage{
geo_create_diagnosisbatch(sample_metadata)
}
\arguments{
\item{sample_metadata}{A data frame. The data frame holding the
sample metadata.}
}
\description{
Helper function to create diagnosis and batch
}
