% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_create_diagnosisdesignbatch.R
\name{geo_create_diagnosisdesignbatch}
\alias{geo_create_diagnosisdesignbatch}
\title{geo_create_diagnosisdesignbatch}
\usage{
geo_create_diagnosisdesignbatch(sample_metadata)
}
\arguments{
\item{sample_metadata}{A data frame. The data frame holding the
sample metadata.}
}
\description{
Helper function to create diagnosis, design and batch
}
