% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_create_expressionset.R
\name{geo_create_expressionset}
\alias{geo_create_expressionset}
\title{geo_create_expressionset}
\usage{
geo_create_expressionset(eset, idtype)
}
\arguments{
\item{eset}{The raw expression set, downloaded with the `GEOquery::getGEO`
funciton.}

\item{idtype}{A character string. The type of ID used to name the
genes. One of 'entrez' or 'affy' intended to use either entrez IDs or
affy IDs. Caution: when using entrez IDs, missing and duplicated IDs
are being removed!}
}
\description{
Helper function to generate a Expression Set.
}
