% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_load_esets.R
\name{geo_load_eset}
\alias{geo_load_eset}
\title{geo_load_eset}
\usage{
geo_load_eset(
  name,
  datadir,
  targetcolname,
  targetlevelname,
  controllevelname,
  targetcol,
  targetname,
  controlname,
  use_rawdata = FALSE,
  setid = 1
)
}
\arguments{
\item{name}{A character string. The name of the GEO study.}

\item{datadir}{A character string. The path to the directory to store
the downloaded GEO datasets.}

\item{targetcolname}{A character string. The name of the column in the
GEO expression set's pheno data, holding the variable of interest
(mapping information).}

\item{targetlevelname}{A character string. The name of the the targets in
`targetcolname` (mapping information).}

\item{controllevelname}{A character string. The name of the the controls in
`targetcolname` (mapping information).}

\item{targetcol}{A character string. The columname of `sample_metadata`
holding the targets (default: "target"; caution: this should not
be changed!).}

\item{targetname}{A character string. Name of the the targets,
specified in the 'target' column of `sample_metadata` (default: "Target";
caution: this should not be changed!).}

\item{controlname}{A character string. Name of the the controls,
specified in the 'target' column of `sample_metadata` (default: "Control";
caution: this should not be changed!).}

\item{use_rawdata}{A logical. Indicates, if raw the data is downloaded
in CEL file format, uncompressed and subsequently normalized with a GCRMA
normalization, or if the already normalized datasets from GEO are
downloaded (default: FALSE).}

\item{setid}{A integer. Indicates, which set should be used by the software
for subsequent analyses (default: 1).}
}
\description{
Helper function to load expression sets from GEO.
}
