% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_geo_data.R
\name{load_geo_data}
\alias{load_geo_data}
\title{Load datasets from the GEO database}
\usage{
load_geo_data(studiesinfo, datadir)
}
\arguments{
\item{studiesinfo}{A list that contains specifications
on the studie's / studies' meta data.}

\item{datadir}{A character string. Path to the data-folder.}
}
\description{
Load datasets that are stored on the GEO databse
}
