% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog_studiesinfo.R
\name{prog_studiesinfo}
\alias{prog_studiesinfo}
\title{prog_studiesinfo}
\usage{
prog_studiesinfo(tab, type)
}
\arguments{
\item{tab}{A table containing the studiesinfo.}

\item{type}{A character string. Either "discovery" or "validation".}
}
\description{
Helper function to import studiesinfos for prognostic
  studies from table
}
\details{
The function takes a table as input and converts it to a
  studiesinfo list.
}
