% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog_studiesinfo.R
\name{progclassifier_studyinfo}
\alias{progclassifier_studyinfo}
\title{progclassifier_studyinfo}
\usage{
progclassifier_studyinfo(tab)
}
\arguments{
\item{tab}{A table containing the studiesinfo.}
}
\description{
Helper function to import studiesinfos for the prognostic
  classifier
}
\details{
The function takes a table as input and converts it to a
  prognostic classifier.
}
