% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studiesinfo_from_table.R
\name{studiesinfo_from_table}
\alias{studiesinfo_from_table}
\title{studiesinfo_from_table}
\usage{
studiesinfo_from_table(tab)
}
\arguments{
\item{tab}{A table containing the studiesinfo.}
}
\description{
Helper function to import studiesinfo from table
}
\details{
The function takes a table as input and converts it to a
  studiesinfo list.
}
