% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_geo_data.R
\name{load_geo_data}
\alias{load_geo_data}
\title{Load datasets from the GEO database}
\usage{
load_geo_data(
  studiesinfo,
  datadir,
  plotdir = "./plots/",
  idtype,
  visualize_batches = TRUE
)
}
\arguments{
\item{studiesinfo}{A list that contains specifications
on the studie's / studies' meta data.}

\item{datadir}{A character string. Path to the data-folder.}

\item{plotdir}{A character string. Path to the folder to store resulting
plots. Default: "./plots/".}

\item{idtype}{A character string. The type of ID used to name the
genes. One of 'entrez' or 'affy' intended to use either entrez IDs or
affy IDs. Caution: when using entrez IDs, missing and duplicated IDs
are being removed!}

\item{visualize_batches}{A logical. Indicates, if batch effects and batch
effect removals should be plotted (default: TRUE).}
}
\description{
Load datasets that are stored on the GEO databse
}
\details{
The functions writes objects to the global environment, including
  the expression sets of the studies specified in `studiesinfo` - the
  here used study names (list keys) are used for the naming in the global
  environment. Furthermore, `mergedset`, a large expression set containing
  all studies, is also written to the global environment. Please note,
  that this set contains the raw expression data with batch effects.
  Bbatch effects are detected, removed and provided with the object
  `mergeset`, a matrix containing batch corrected expression data with
  genes in the rows and samples in the columns. `sample_metadata` is a
  data.frame, which holds information on the samples which are included
  in the studies, including if they are a "Target" or a "Control".
  `diagnosis` (a binary coding of the target variable) and
  `batch` () are also provided to the global environment to be used with
  other functions of the `sigident` R package.
}
\references{
\url{https://gitlab.miracum.org/clearly/sigident}
}
\seealso{
\code{sigident}
}
